refactor : validation error throw 하는 방식으로 수정, 유저 삭제시 alert 추가
