+++ src/js/utils/confirm.js
@@ -1,0 +1,6 @@
+const confirmAlert = (message) => {
+  const ans = confirm(message);
+  return ans;
+};
+
+export default confirmAlert;
