--- src/js/utils/messages.js
@@ -1,3 +1,0 @@
-export const invaildUserName = '이름은 2글자 이상이어야 해요!';
-export const invalidTodo = '2글자 이상 입력해주세요!';
-export const defaultErrorMessage = '잠시후에 다시 시도해주세요';
